fix(webui): move client wrapper outside generated client-ts dir
