+++ webui/scripts/verify-client-wrapper-boundary.sh
@@ -16,6 +16,9 @@ if [ "$#" -gt 0 ]; then
       src/client-ts/*)
         continue
         ;;
+      src/lib/clientWrapper.ts|src/lib/clientWrapper.test.ts)
+        continue
+        ;;
       *.ts|*.tsx|*.js|*.jsx)
         if [ -f "$file" ]; then
           scan_targets+=("$file")
@@ -29,29 +32,36 @@ if [ "${#scan_targets[@]}" -gt 0 ]; then
   if command -v rg >/dev/null 2>&1; then
     if rg -n -- "$PATTERN" "${scan_targets[@]}"; then
       echo "❌ Direct client-ts/*.gen import detected outside wrapper boundary."
-      echo "   Use imports from src/client-ts (index) or src/client-ts/wrapper."
+      echo "   Use imports from src/client-ts (index) or src/lib/clientWrapper."
       exit 1
     fi
   else
     if grep -n -E "$PATTERN" "${scan_targets[@]}"; then
       echo "❌ Direct client-ts/*.gen import detected outside wrapper boundary."
-      echo "   Use imports from src/client-ts (index) or src/client-ts/wrapper."
+      echo "   Use imports from src/client-ts (index) or src/lib/clientWrapper."
       exit 1
     fi
   fi
 else
   if command -v rg >/dev/null 2>&1; then
-    if rg -n --glob '*.{ts,tsx,js,jsx}' --glob '!src/client-ts/**' -- "$PATTERN" src tests; then
+    if rg -n \
+      --glob '*.{ts,tsx,js,jsx}' \
+      --glob '!src/client-ts/**' \
+      --glob '!src/lib/clientWrapper.ts' \
+      --glob '!src/lib/clientWrapper.test.ts' \
+      -- "$PATTERN" src tests; then
       echo "❌ Direct client-ts/*.gen import detected outside wrapper boundary."
-      echo "   Use imports from src/client-ts (index) or src/client-ts/wrapper."
+      echo "   Use imports from src/client-ts (index) or src/lib/clientWrapper."
       exit 1
     fi
   else
     if grep -R -n -E "$PATTERN" src tests \
       --include='*.ts' --include='*.tsx' --include='*.js' --include='*.jsx' \
-      --exclude-dir='client-ts'; then
+      --exclude-dir='client-ts' \
+      --exclude='clientWrapper.ts' \
+      --exclude='clientWrapper.test.ts'; then
       echo "❌ Direct client-ts/*.gen import detected outside wrapper boundary."
-      echo "   Use imports from src/client-ts (index) or src/client-ts/wrapper."
+      echo "   Use imports from src/client-ts (index) or src/lib/clientWrapper."
       exit 1
     fi
   fi
